Adminlte emiliano

See merge request codigitar/setecel!3

+++ app/Http/Livewire/Usuarios.php
@@ -4,11 +4,14 @@ namespace App\Http\Livewire;
 use App\Models\User;
 use App\Models\Rol;
 use Livewire\Component;
+use Illuminate\Http\Request;
 
 class Usuarios extends Component
 {
-    public  $idu, $name, $email, $nombre_y_apellido, $telefono, $dni, $activo, $domicilio, $users, $userup,$roles, $search;
-    public $funcion, $funcionru;
+    
+    public $idus, $usuarios, $idu, $name, $email, $nombre_y_apellido, $telefono, $dni, $activo, $domicilio, $users, $userup,$roles,$roless, $search;
+    public $funcion="", $funcionru, $userlog;
+    
     public function render()
     {
         $this->roles = Rol::where('nombre','LIKE','%' . $this->search . '%')
@@ -27,24 +30,49 @@ class Usuarios extends Component
    
     public function store()
     {
-        User::create([
-            'name' => $this->name,
-            'email' => $this->email,
-            'nombre_y_apellido'=>$this->nombre_y_apellido,
-            'domicilio'=>$this->domicilio,
-            'telefono'=>$this->telefono,
-            'dni'=>$this->dni,
-        ]);
+        
+        if (auth()->user()->cannot('store', auth()->user())) {
+            abort(403);
+        }else
+        {
+            User::create([
+                'name' => $this->name,
+                'email' => $this->email,
+                'nombre_y_apellido'=>$this->nombre_y_apellido,
+                'domicilio'=>$this->domicilio,
+                'telefono'=>$this->telefono,
+                'dni'=>$this->dni,
+            ]);
+            $this->funcion="";
+        }
     }
 
     public function destruir(User $user)
     {
-        $user->delete();
+        if (auth()->user()->cannot('destruir', auth()->user())) {
+            abort(403);
+        }else
+        {
+            $user->delete();
+        }
     }
 
     public function funcion()
     {
         $this->funcion="crear";
+        $this->funcionru="";
+        $this->idu=null;
+        $this->name=null;
+        $this->nombre_y_apellido=null;
+        $this->domicilio=null;
+        $this->telefono=null;
+        $this->email=null;
+        $this->dni=null;
+    }
+    
+    public function endfunctions()
+    {
+        $this->funcion="";
     }
 
     public function update(User $user)
@@ -57,25 +85,41 @@ class Usuarios extends Component
         $this->email=$user->email;
         $this->dni=$user->dni;
         $this->funcion="adaptar";
+        $this->funcionru="";
     }
 
     public function editar()
     {
         $userup =User::find($this->idu);
-        
         $userup->nombre_y_apellido=$this->nombre_y_apellido;
         $userup->dni=$this->dni;
         $userup->domicilio=$this->domicilio;
         $userup->telefono=$this->telefono;
         $userup->email=$this->email;
         $userup->save();
+        $this->funcion="";
     }
 
     public function rolusuario(User $user)
-    {
+    {   
+        $this->idus=$user->id;
         $this->nombre_y_apellido=$user->nombre_y_apellido;
         $this->funcionru="asigna";
+        $this->funcion="";   
     }
 
+    public function asignarols(Rol $rol)        
+    {      
+        $this->roless= User::find($this->idus)->rols()->where('rol_id',$rol->id)->get(); 
+        if(sizeof($this->roless)==0)
+        {
+            $rol->users()->attach($this->idus);
+        }            
+    }
 
+    public function quitarol(Rol $rol)
+    {   
+        
+        $rol->users()->detach($this->idus);                      
+    }
 }
